pyproject.toml
[tool.poetry]
name = "django_inertia"
version = "0.1.0"
description = "Django + Inertia.js test app"
authors = ["Cristian Rasch <cristianrasch@fastmail.fm>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.4"
honcho = "^1.0.1"
django-js-routes = "^0.1.4"
inertia-django = "^0.3.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
